requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
instagram-private-api==1.6.0.0
numpy==1.18.4
opencv-python==4.2.0.34
python-dotenv==0.13.0
ratelimiter==1.2.0.post0
requests==2.23.0
typing==3.7.4.1
urllib3==1.25.9
